Add CI + test-settings shim + split tests into a package (#1279)
First PR off the testing roadmap (#1278). Three interdependent infra
changes; no production code or test logic changes.

1. Test-settings shim — makeabilitylab/settings_test.py sets
   MIGRATION_MODULES={'website': None} so the test DB is built directly
   from the current models, sidestepping the gitignored, per-environment
   website/migrations/ history. Durable fix for the "column already exists"
   test-DB flakiness (#1267). Also reads DATABASE_HOST/PORT from env so CI
   can point at its Postgres service.

2. GitHub Actions CI — .github/workflows/test.yml runs the suite on every
   push to master and every PR, against a postgres:16 service container,
   mirroring the Dockerfile's ImageMagick/Ghostscript deps. Reports status
   only (free/unlimited for this public repo); does not block pushes.

3. Split website/tests.py (~1,580 lines) into a website/tests/ package:
   tests/base.py (DatabaseTestCase + fixtures) plus test_*.py by concern.
   Pure move — 99 tests before and after, all passing.

Docs (CLAUDE.md, CONTRIBUTING.md, docs/DEPLOYMENT.md) updated to use the
shim and document CI.

Verified: `manage.py test website --settings=makeabilitylab.settings_test`
runs 99 tests OK on a clean test DB with no manual DROP DATABASE step.
